Add a [P]in action to save a hashtag timeline
This goes along with the ability to "go to" a saved hashtag timeline.

We need to pass the "local" flag down build_timeline() method so that
this is also saved. Note that we cannot save both a local and a
federated timeline for the same hashtag: I'm not sure it's worth
supporting both options?