Prevent toot timeline to break when piped to extenal programs
Toot timeline try to ask the user to continue in the next page. This
dont works if the command is piped to any other software.

We stop the query in the first loop when not in a tty context. The
command should be called with the expected item count in this case.