Implement comment metadata support in editor
This allows setting some properties of the toot directly while editing
the message. It is backwards-compatible with the use of # for comment
lines and forwards-compatible with the introduction of new metadata
comments in the future. The supported ones are listed in the explanatory
message appended to the text in the editor.