Remove comment support and switch to YAML-style header metadata
As discussed on IRC, remove the filtering of lines starting
with a hash-sign. This allows pasting snippets of code that with # as
comment symbol, and allows hashtags to start on an empy line.

Metadata support is moved to a YAML-style header similar to what Jekyll
uses. We don't go full YAML though, just a simple key: value syntax
(the same as before, but without the # in front).