Fix for issue #217 Custom LOCAL_DOMAIN instance setting isn't used
Switched api get_instance to use v2 api, falling back to v1 if
v2 not available.

v2 api returns a "domain" attribute; we use this if available
in preference to app.instance when filling in local account
names' domains. This fixes display of users on servers that
have a custom LOCAL_DOMAIN setting.  It also fixes retrieval
of those users when we show accounts.