Move to Python 3.7 as a minimum version
The term-image library, needed for image display, is Python 3.7+
so I have made what I think are the minimum set of changes to
move to Python 3.7 - testing with 3.7+ only, as a start.

Unchanged- vermin target is still 3.6 because changing it to 3.7
fails for an unknown reason. Seems harmless to leave it at 3.6
for the time being.